Ver. 2.0
I've done some more tweaking to the latency test. UI elements have been added as well as the ability to move back from the second test to the first. At the end of the second test, both latency values are displayed. I will work on this further to add better functionality and documentation