Logic placement and key detection
As of now, the ScoreCheck script should find the keys being pressed at a given time and compare them against the keys being used for the game. Then said keys are checked as to whether or not they match a note to be hit at that time and if so, calculate the necessary scoring accordingly.