Blank MIDI merging
Current state of testing. Created a blank MIDI file of one measure in 4/4 time at 120 BPM. This MIDI is in the assets folder as "blank.mid". Using the MergeSequentially method in DryWetMidi, I merge the blank midi with the midi to be played in order to offset the delay lag. This appears to work