Implementation of Instructions and Countdown
The main game scene has been updated to display the instructions to playing the game only the first time the game is ever played. Once that is cleared and/or the game is loaded any time after that, a three second countdown will be displayed and prevent all gameplay until the countdown is over.